refactor(SCIM): rename sso/SSOTab.tsx to sso/index.tsx
Mirror the `general-tab/index.tsx` pattern used elsewhere in
`organisation-settings/tabs/`. The component is the SSO folder's
entry point, not a standalone tab file.
